Fixed the execution time been poorly calculated and abstracted the time format function
